commenting downlink and radio control parts

+++ sw/airborne/math/Int32Vect2.java
@@ -1,8 +1,8 @@
 package sw.airborne.math;
 
 public class Int32Vect2 {
-	public Long x;
-	public Long y;
+	public int x;
+	public int y;
 	public boolean equals(Int32Vect2 _a){
 		if(this.x==_a.x) return false;
 		if(this.y==_a.y) return false;
