+++ sw/airborne/math/NedCoor_i.java
@@ -1,7 +1,7 @@
 package sw.airborne.math;
 
 public class NedCoor_i{
-	public Long x, y, z;
+	public int x, y, z;
 	
 	public boolean notequals0(){
 		if(x!=0 && y!= 0 & z!= 0) return true;
